requirements.txt
boto3~=1.17
botocore~=1.20
clickhouse_driver~=0.2.0
cloudpickle~=2.0
ctgan~=0.5.1
deepecho~=0.3.0.post1
matplotlib~=3.4
numpy~=1.21
pandas~=1.1
pyarrow~=7.0
pytorch_lightning~=1.3
rdt~=0.6.3
scikit_learn~=0.24
scipy~=1.6
sdv~=0.13.1
SQLAlchemy~=1.3.24
tensorflow~=2.8
torchmetrics~=0.3
tqdm~=4.63
vaex~=4.7
plotly~=5.6